refactor: rewrite in ES6 notation task
